+++ test/asset/outboxProject/srv/service/serviceCustomHooks.js
@@ -1,0 +1,122 @@
+"use strict";
+
+const cds = require("@sap/cds");
+
+const ACTION = [
+  "action",
+  "main",
+  "anotherAction",
+  "actionClusterByPayloadWithCb",
+  "actionClusterByPayloadWithoutCb",
+  "actionClusterByEventWithCb",
+  "actionClusterByDataWithCb",
+  "actionClusterByEventWithoutCb",
+  "actionClusterByData",
+];
+
+class OutboxCustomHooks extends cds.Service {
+  async init() {
+    await super.init();
+
+    for (const actionName of ACTION) {
+      this.on(actionName, (req) => {
+        cds.log(this.name).info(req.event, {
+          data: req.data,
+          user: req.user.id,
+          subType: req.eventQueue.processor.eventSubType,
+        });
+      });
+    }
+
+    this.on("clusterQueueEntries.action", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByPayloadProperty("event");
+    });
+
+    this.on("clusterQueueEntries.actionClusterByData", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByPayloadProperty("data.to");
+    });
+
+    this.on("clusterQueueEntries.actionClusterByDataWithCb", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByDataProperty("to", (clusterKey, entries) => {
+        return { ...entries[0], guids: entries.map((a) => a.guids).flat() };
+      });
+    });
+
+    this.on("clusterQueueEntries", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByPayloadProperty("data.to");
+    });
+
+    this.on("clusterQueueEntries.actionClusterByPayloadWithCb", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByPayloadProperty("data.to", (clusterKey, entries) => {
+        return { ...entries[0], guids: entries.map((a) => a.guids).flat() };
+      });
+    });
+
+    this.on("clusterQueueEntries.actionClusterByPayloadWithoutCb", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByPayloadProperty("data.to");
+    });
+
+    this.on("clusterQueueEntries.actionClusterByEventWithCb", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByEventProperty("referenceEntityKey", (clusterKey, entries) => {
+        return { ...entries[0], guids: entries.map((a) => a.guids).flat() };
+      });
+    });
+
+    this.on("clusterQueueEntries.actionClusterByEventWithoutCb", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+      });
+      return req.eventQueue.clusterByEventProperty("referenceEntityKey");
+    });
+
+    this.on("checkEventAndGeneratePayload", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+        subType: req.eventQueue.processor.eventSubType,
+      });
+      return req.data;
+    });
+
+    this.on("checkEventAndGeneratePayload.action", (req) => {
+      cds.log(this.name).info(req.event, {
+        data: req.data,
+        user: req.user.id,
+        subType: req.eventQueue.processor.eventSubType,
+      });
+      req.data.to = "newValue";
+      return req.data;
+    });
+  }
+}
+
+module.exports = OutboxCustomHooks;